history_study/requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
coverage==6.5.0
cryptography==36.0.2
cssselect==1.1.0
Deprecated==1.2.13
distlib==0.3.6
Django==4.0.6
docx==0.2.4
exceptiongroup==1.0.0rc9
fake-useragent==0.1.11
filelock==3.8.0
flake8==5.0.4
h11==0.14.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.7.0
itemloaders==1.0.5
jmespath==1.0.1
lxml==4.9.1
mccabe==0.7.0
mock==4.0.3
outcome==1.2.0
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
Protego==0.2.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
PyDispatcher==2.0.5
pyflakes==2.5.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-cov==4.0.0
python-docx==0.8.11
queuelib==1.6.2
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
scrapy-redis==0.7.3
selenium==4.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.1
tomli==2.0.1
tox==3.26.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.4.0
typing-extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.5
w3lib==2.0.1
wrapt==1.14.1
wsproto==1.2.0
zope.interface==5.4.0
